style (extract main style definitions and moved them to to component stylesheets)
