+++ src/app/view/modules/other/progress-indicator/progress-indicator.component.css
@@ -1,0 +1,44 @@
+salsah-progress-indicator, .progress-indicator {
+  width: 96px;
+  height: 56px;
+  padding: 24px 36px 24px 36px; }
+  salsah-progress-indicator.page-center, .progress-indicator.page-center {
+    position: absolute;
+    top: 39%;
+    left: 50%;
+    transform: translate(-50%, -50%); }
+  salsah-progress-indicator p, salsah-progress-indicator h1, salsah-progress-indicator h2, salsah-progress-indicator h3, .progress-indicator p, .progress-indicator h1, .progress-indicator h2, .progress-indicator h3 {
+    text-align: center;
+    color: #555555; }
+  salsah-progress-indicator .line, .progress-indicator .line {
+    margin: 0 auto 0;
+    width: 70px;
+    text-align: center; }
+  salsah-progress-indicator .line > div, .progress-indicator .line > div {
+    width: 18px;
+    height: 18px;
+    background-color: #00695c;
+    border-radius: 6px;
+    display: inline-block;
+    -webkit-animation: sk-bouncedelay 1.4s infinite ease-in-out both;
+    animation: sk-bouncedelay 1.4s infinite ease-in-out both; }
+  salsah-progress-indicator .line .bounce1, .progress-indicator .line .bounce1 {
+    -webkit-animation-delay: -0.32s;
+    animation-delay: -0.32s; }
+  salsah-progress-indicator .line .bounce2, .progress-indicator .line .bounce2 {
+    -webkit-animation-delay: -0.16s;
+    animation-delay: -0.16s; }
+@-webkit-keyframes sk-bouncedelay {
+  0%, 80%, 100% {
+    -webkit-transform: scale(0); }
+  40% {
+    -webkit-transform: scale(1); } }
+@keyframes sk-bouncedelay {
+  0%, 80%, 100% {
+    -webkit-transform: scale(0);
+    transform: scale(0); }
+  40% {
+    -webkit-transform: scale(1);
+    transform: scale(1); } }
+
+/*# sourceMappingURL=progress-indicator.component.css.map */
