+++ src/app/view/modules/other/progress-indicator/progress-indicator.component.scss
@@ -1,0 +1,64 @@
+@import '../../../../../assets/style/config';
+
+//
+// two lines with three bouncing squares each
+//
+
+salsah-progress-indicator, .progress-indicator {
+
+    width: 96px;
+    height: 56px;
+    padding: 24px 36px 24px 36px;
+
+    &.page-center {
+        position: absolute;
+        top: 39%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+    }
+    p, h1, h2, h3 {
+        text-align: center;
+        color: #555555;
+    }
+    .line {
+        margin: 0 auto 0;
+        width: 70px;
+        text-align: center;
+    }
+
+    .line > div {
+        width: 18px;
+        height: 18px;
+        background-color: $primary;
+
+        border-radius: 6px;
+        display: inline-block;
+        -webkit-animation: sk-bouncedelay 1.4s infinite ease-in-out both;
+        animation: sk-bouncedelay 1.4s infinite ease-in-out both;
+    }
+
+    .line .bounce1 {
+        -webkit-animation-delay: -0.32s;
+        animation-delay: -0.32s;
+    }
+
+    .line .bounce2 {
+        -webkit-animation-delay: -0.16s;
+        animation-delay: -0.16s;
+    }
+
+    @-webkit-keyframes sk-bouncedelay {
+        0%, 80%, 100% { -webkit-transform: scale(0) }
+        40% { -webkit-transform: scale(1.0) }
+    }
+
+    @keyframes sk-bouncedelay {
+        0%, 80%, 100% {
+            -webkit-transform: scale(0);
+            transform: scale(0);
+        } 40% {
+              -webkit-transform: scale(1.0);
+              transform: scale(1.0);
+          }
+    }
+}
