+++ src/app/view/modules/other/progress-indicator/progress-indicator.component.ts
@@ -10,6 +10,7 @@ export class ProgressIndicatorComponent implements OnInit {
   constructor() { }
 
   ngOnInit() {
+
   }
 
 }
