+++ src/app/view/modules/other/progress-stepper/progress-stepper.component.css
@@ -1,0 +1,46 @@
+.progress-wrapper {
+  width: 100%; }
+  .progress-wrapper .progress-stepper {
+    flex-direction: row;
+    align-items: center;
+    white-space: nowrap;
+    box-sizing: border-box;
+    margin: -24px -24px -20px -24px;
+    padding: 0 24px;
+    border-bottom: 2px solid rgba(0, 0, 0, 0.1);
+    background-color: whitesmoke; }
+    .progress-wrapper .progress-stepper p {
+      width: 100%;
+      display: inline-flex; }
+      .progress-wrapper .progress-stepper p span:nth-child(even) {
+        flex: 1 1 auto;
+        margin-top: 6px; }
+      .progress-wrapper .progress-stepper p span.inactive {
+        color: #7a7a7a; }
+    .progress-wrapper .progress-stepper p.step-number span:nth-child(odd) {
+      padding: 4px;
+      border-radius: 50%;
+      width: 24px;
+      height: 24px;
+      text-align: center; }
+      .progress-wrapper .progress-stepper p.step-number span:nth-child(odd).active {
+        background-color: #ffc400;
+        border: 1px solid #7a7a7a; }
+      .progress-wrapper .progress-stepper p.step-number span:nth-child(odd).done {
+        background-color: #00695c;
+        color: #f9f9f9; }
+    .progress-wrapper .progress-stepper p.step-description {
+      margin-top: -16px; }
+      .progress-wrapper .progress-stepper p.step-description span label {
+        text-align: center;
+        font-weight: 400;
+        font-size: small; }
+      .progress-wrapper .progress-stepper p.step-description span.active label {
+        font-weight: 700; }
+    .progress-wrapper .progress-stepper.error {
+      background-color: indianred;
+      font-weight: 400;
+      font-size: smaller;
+      padding: 12px; }
+
+/*# sourceMappingURL=progress-stepper.component.css.map */
