+++ src/app/view/modules/other/progress-stepper/progress-stepper.component.scss
@@ -1,0 +1,72 @@
+@import '../../../../../assets/style/config';
+
+.progress-wrapper {
+    width: 100%;
+
+    .progress-stepper {
+        flex-direction: row;
+        align-items: center;
+        white-space: nowrap;
+        box-sizing: border-box;
+        margin: -24px -24px -20px -24px;
+        padding: 0 24px;
+        border-bottom: 2px solid rgba(0, 0, 0, 0.1);
+        background-color: whitesmoke;
+
+        p {
+            width: 100%;
+            display: inline-flex;
+            span {
+                &:nth-child(even) {
+                    flex: 1 1 auto;
+                    margin-top: 6px;
+                }
+                &.inactive {
+                    color: darken($bright, 50%);
+                }
+            }
+        }
+
+        p.step-number {
+            span {
+                &:nth-child(odd) {
+                    padding: 4px;
+                    border-radius: 50%;
+                    width: 24px;
+                    height: 24px;
+                    text-align: center;
+
+                    &.active {
+                        background-color: $accent;
+                        //                        @include md-box-shadow();
+                        border: 1px solid darken($bright, 50%);
+                    }
+
+                    &.done {
+                        background-color: $primary;
+                        color: $bright;
+                    }
+                }
+            }
+        }
+
+        p.step-description {
+            margin-top: -16px;
+            span label {
+                text-align: center;
+                font-weight: 400;
+                font-size: small;
+            }
+            span.active label {
+                font-weight: 700;
+            }
+        }
+
+        &.error {
+            background-color: indianred;
+            font-weight: 400;
+            font-size: smaller;
+            padding: 12px;
+        }
+    }
+}
