+++ src/assets/style/_progress.scss
@@ -1,131 +1,1 @@
-/* progress indicator
- * two lines with three bouncing squares each
- *
- */
 
-salsah-progress-indicator, .progress-indicator {
-
-    width: 96px;
-    height: 56px;
-    padding: 24px 36px 24px 36px;
-
-    &.page-center {
-        position: absolute;
-        top: 39%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-    }
-    p, h1, h2, h3 {
-        text-align: center;
-        color: #555555;
-    }
-    .line {
-        margin: 0 auto 0;
-        width: 70px;
-        text-align: center;
-    }
-
-    .line > div {
-        width: 18px;
-        height: 18px;
-        background-color: rgba(0, 105, 92, 1);
-
-        border-radius: 6px;
-        display: inline-block;
-        -webkit-animation: sk-bouncedelay 1.4s infinite ease-in-out both;
-        animation: sk-bouncedelay 1.4s infinite ease-in-out both;
-    }
-
-    .line .bounce1 {
-        -webkit-animation-delay: -0.32s;
-        animation-delay: -0.32s;
-    }
-
-    .line .bounce2 {
-        -webkit-animation-delay: -0.16s;
-        animation-delay: -0.16s;
-    }
-
-    @-webkit-keyframes sk-bouncedelay {
-        0%, 80%, 100% { -webkit-transform: scale(0) }
-        40% { -webkit-transform: scale(1.0) }
-    }
-
-    @keyframes sk-bouncedelay {
-        0%, 80%, 100% {
-            -webkit-transform: scale(0);
-            transform: scale(0);
-        } 40% {
-              -webkit-transform: scale(1.0);
-              transform: scale(1.0);
-          }
-    }
-}
-
-
-
-salsah-progress-stepper{
-    width: 100%;
-    div.salsah-progress-stepper {
-        flex-direction: row;
-        align-items: center;
-        white-space: nowrap;
-        box-sizing: border-box;
-        margin: -24px -24px -20px -24px;
-        padding: 0 24px;
-        border-bottom: 2px solid rgba(0,0,0,0.1);
-        background-color: whitesmoke;
-        p {
-            width: 100%;
-            display: inline-flex;
-            span {
-                &:nth-child(even) {
-                    flex: 1 1 auto;
-                    margin-top: 6px;
-                }
-                &.inactive {
-                    color: darken($bright, 50%);
-                }
-            }
-        }
-        p.step-number {
-            span {
-                &:nth-child(odd) {
-                    padding:4px;
-                    border-radius: 50%;
-                    width: 24px;
-                    height: 24px;
-                    text-align: center;
-
-                    &.active {
-                        background-color: $accent;
-//                        @include md-box-shadow();
-                        border: 1px solid darken($bright, 50%);
-                    }
-
-                    &.done {
-                        background-color: $primary;
-                        color: $bright;
-                    }
-                }
-            }
-        }
-        p.step-description {
-            margin-top: -16px;
-            span label {
-                text-align: center;
-                font-weight: 400;
-                font-size: small;
-            }
-            span.active label {
-                font-weight: 700;
-            }
-        }
-        &.error {
-            background-color: indianred;
-            font-weight: 400;
-            font-size: smaller;
-            padding: 12px;
-        }
-    }
-}
