+++ src/assets/style/main.scss
@@ -11,10 +11,10 @@
 // You should have received a copy of the GNU Affero General Public
 // License along with SALSAH.  If not, see <http://www.gnu.org/licenses/>.
 //
+
 @import 'mixins';
 @import 'config';
 @import 'error';
-@import 'progress';
 @import 'framework';
 @import 'listing';
 @import 'form';
