+++ src/styles.css
@@ -9,4 +9,4 @@
 @import './assets/style/md-teal-amber.min.css';
 
 /* own style setting */
-@import './assets/style/main.min.css';
+@import './assets/style/main.css';
